Conflicts:
	apps/spaintgui/RiftRenderer.cpp
	apps/spaintgui/main.cpp

+++ modules/spaint/src/core/SpaintPipeline.cpp
@@ -246,7 +246,7 @@ void SpaintPipeline::run_training_section(const RenderState_CPtr& samplingRender
   LabelManager_CPtr labelManager = m_model->get_label_manager();
   const size_t maxLabelCount = labelManager->get_max_label_count();
   bool *labelMask = m_labelMaskMB->GetData(MEMORYDEVICE_CPU);
-  for(int i = 0; i < maxLabelCount; ++i)
+  for(size_t i = 0; i < maxLabelCount; ++i)
   {
     labelMask[i] = labelManager->has_label(static_cast<SpaintVoxel::LabelType>(i));
   }
@@ -257,7 +257,7 @@ void SpaintPipeline::run_training_section(const RenderState_CPtr& samplingRender
   m_voxelSampler->sample_voxels(raycastResult, m_model->get_scene().get(), *m_labelMaskMB, *m_sampledVoxelsMB, *m_sampledVoxelCountsMB);
 
   // TEMPORARY: Output the numbers of voxels sampled for each label (for debugging purposes).
-  for(int i = 0; i < m_sampledVoxelCountsMB->dataSize; ++i)
+  for(size_t i = 0; i < m_sampledVoxelCountsMB->dataSize; ++i)
   {
     std::cout << m_sampledVoxelCountsMB->GetData(MEMORYDEVICE_CPU)[i] << ' ';
   }
